+++ core/src/main/java/arez/ComputedValueInfoImpl.java
@@ -4,6 +4,7 @@ import arez.spy.ComponentInfo;
 import arez.spy.ComputedValueInfo;
 import arez.spy.ObservableValueInfo;
 import arez.spy.ObserverInfo;
+import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
@@ -11,6 +12,7 @@ import java.util.Objects;
 import java.util.stream.Collectors;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import static org.realityforge.braincheck.Guards.*;
 
 /**
  * A implementation of {@link ComputedValueInfo} that proxies to a {@link ComputedValue}.
@@ -86,7 +88,25 @@ final class ComputedValueInfoImpl
   @Override
   public List<ObservableValueInfo> getDependencies()
   {
-    return _spy.getDependencies( _computedValue );
+    if ( _computedValue.isComputing() )
+    {
+      final Transaction transaction = getTransactionComputing();
+      final ArrayList<ObservableValue<?>> observableValues = transaction.getObservableValues();
+      if ( null == observableValues )
+      {
+        return Collections.emptyList();
+      }
+      else
+      {
+        // Copy the list removing any duplicates that may exist.
+        final List<ObservableValue<?>> list = observableValues.stream().distinct().collect( Collectors.toList() );
+        return ObservableValueInfoImpl.asUnmodifiableInfos( list );
+      }
+    }
+    else
+    {
+      return ObservableValueInfoImpl.asUnmodifiableInfos( _computedValue.getObserver().getDependencies() );
+    }
   }
 
   /**
@@ -160,5 +180,37 @@ final class ComputedValueInfoImpl
   public int hashCode()
   {
     return _computedValue.hashCode();
+  }
+
+  /**
+   * Return the transaction that is computing specified ComputedValue.
+   */
+  @Nonnull
+  Transaction getTransactionComputing()
+  {
+    assert _computedValue.isComputing();
+    final Transaction transaction = getTrackerTransaction( _computedValue.getObserver() );
+    if ( Arez.shouldCheckInvariants() )
+    {
+      invariant( () -> transaction != null,
+                 () -> "Arez-0106: ComputedValue named '" + _computedValue.getName() + "' is marked as computing but " +
+                       "unable to locate transaction responsible for computing ComputedValue" );
+    }
+    assert null != transaction;
+    return transaction;
+  }
+
+  /**
+   * Get transaction with specified observer as tracker.
+   */
+  @Nullable
+  private Transaction getTrackerTransaction( @Nonnull final Observer observer )
+  {
+    Transaction t = _computedValue.getContext().getTransaction();
+    while ( null != t && t.getTracker() != observer )
+    {
+      t = t.getPrevious();
+    }
+    return t;
   }
 }