+++ core/src/main/java/arez/Spy.java
@@ -82,19 +82,6 @@ public interface Spy
   List<ObserverInfo> getObservers( @Nonnull ComputedValue<?> computedValue );
 
   /**
-   * Return the list of dependencies of the ComputedValue.
-   * The list is an immutable copy of the dependencies of the {@link ComputedValue}.
-   * If the {@link ComputedValue} is currently being computed (i.e. {@link #isComputing(ComputedValue)}
-   * returns true) then the dependencies are provisional and may be added to as transaction
-   * completes.
-   *
-   * @param computedValue the ComputedValue.
-   * @return the list of dependencies for ComputedValue.
-   */
-  @Nonnull
-  List<ObservableValueInfo> getDependencies( @Nonnull ComputedValue<?> computedValue );
-
-  /**
    * Return true if the ObservableValue is a ComputedValue.
    *
    * @param observableValue the ObservableValue.